tests: add case to compare_signed.bats for v2.5 style releases
To match the current Linux (upto 6.3 at least) kernel layout for
IPC4 firmware naming, the firmware files for SOF2.5 will all
have same basename and this is not compatible with how compare_signed.bats
is done.

As this only affects a subset of Intel hardare (the cAVS2.5 platforms),
skip this test and do the comparison checks manually when releasing
new binaries.

Signed-off-by: Kai Vehmanen <kai.vehmanen@linux.intel.com>