Update sof-v2.10 binaries for Intel hardware
Update the community signed binaries for Intel systems. These binaries
are functionally identical, but the original version did not have the
v2.10 git tag in work area, so the version string literals embedded in
the binaries are missing the tag. This causes a binary mismatch between
differently signed binaries of the same version. Fix this mistake for
community binaries.

Community binaries build identification "Build #4714".

Signed-off-by: Kai Vehmanen <kai.vehmanen@linux.intel.com>